Interfaces repository y Service, Implementando inyeccion de dependencias (@Autowired y @Component)
