+++ src/main/java/com/matias/springboot/di/app/springbootdi/controllers/SomeController.java
@@ -2,19 +2,21 @@ package com.matias.springboot.di.app.springbootdi.controllers;
 
 import java.util.List;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.matias.springboot.di.app.springbootdi.models.Product;
-import com.matias.springboot.di.app.springbootdi.services.ProductService;
+import com.matias.springboot.di.app.springbootdi.services.ProductServiceImpl;
 
 @RestController
 @RequestMapping("/api")
 public class SomeController {
 
-    private ProductService service = new ProductService();
+    @Autowired
+    private ProductServiceImpl service;
 
     @GetMapping
     public List<Product> list() {
