+++ src/main/java/com/matias/springboot/di/app/springbootdi/services/ProductServiceImpl.java
@@ -1,0 +1,35 @@
+package com.matias.springboot.di.app.springbootdi.services;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+
+import com.matias.springboot.di.app.springbootdi.models.Product;
+import com.matias.springboot.di.app.springbootdi.repositories.ProductRepositoryImpl;
+
+@Component
+public class ProductServiceImpl implements ProductService {
+
+    @Autowired
+    private ProductRepositoryImpl repository;
+
+    @Override
+    public List<Product> findAll() {
+        return repository.findAll().stream().map(p -> {
+            Double priceTax = p.getPrice() * 1.25d;
+            // Product newProd = new Product(p.getId(), p.getName(), priceTax.longValue());
+            // return newProd;
+            Product newProd = (Product) p.clone();
+            newProd.setPrice(priceTax.longValue());
+            return newProd;
+        }).collect(Collectors.toList());
+    }
+
+    @Override
+    public Product findByiD(Long id) {
+        return repository.findById(id);
+    }
+
+}
